Big change
After discovering that my current architecture was not
compatible with easily using Angular animations, I
basically rewired the whole thing. Now the components
are much leaner with the company service doing more of
the heavy lifting. Search to search results will now
be in the browser history and the first parts of route
animations have been added.