Desktop search focus no longer shifts
On mobile, the search box shift up on focus so that it stays
at the top of the screen and avoids collision with the mobile
keypad. On Desktop however, this is not needed. I created an
AppService to check screensize when the app loads. I will place
other functions that need to be called when the app is loaded.